Enable iOS build target
iOS was disabled due to the fact that the test suite accessed files on the local file system which accessible on the simulator. This commit moves all such tests to a new source set: `fileBasedTest`. All test source set except `ios` depend on this new source set, so running `gradle test` on any platform will run all tests that run on that platform.

The resources for the performance test were moved into `jvmTest` instead of `fileBasedTest`, since the performance test is jvm-only.