Enable JS IR compilation support
Commit 47d352a7 previously disabled IR compilation due to tests not
functioning. In order to use `js(BOTH)`, I removed the unused dependency
on kotlinx-nodejs (that library is just headers based on TypeScript
definitions, it isn't necessary for running on node), and the
unnecessary `binaries.executable()` (it's not necessary for tests, and
there are no other js binaries in this repo).